+++ packages/substrate-bindings/src/codecs/metadata/index.ts
@@ -1,3 +1,3 @@
-export { metadata } from "./metadata"
+export { metadata, decAnyMetadata } from "./metadata"
 export * from "./v14"
 export * from "./v15"